Move arm interface impl into arm arch module
